+++ guillotina/async.py
@@ -5,6 +5,7 @@ from guillotina import logger
 from guillotina.browser import ErrorResponse
 from guillotina.browser import UnauthorizedResponse
 from guillotina.browser import View
+from guillotina.db.transaction import Status
 from guillotina.exceptions import Unauthorized
 from guillotina.transactions import get_tm
 from guillotina.transactions import get_transaction
@@ -42,25 +43,32 @@ class QueueUtility(object):
                 priority, view = await self._queue.get()
                 got_obj = True
                 txn = get_transaction(view.request)
-                if txn is None:
-                    tm = get_tm(view.request)
-                    txn = tm.begin(view.request)
+                tm = get_tm(view.request)
+                if txn is None or txn.status in (Status.ABORTED, Status.COMMITTED):
+                    txn = await tm.begin(view.request)
+                else:
+                    # still finishing current transaction, this connection
+                    # will be cut off, so we need to wait until we no longer
+                    # have an active transaction on the reqeust...
+                    await self.add(view)
+                    await asyncio.sleep(1)
+                    continue
 
                 try:
                     view_result = await view()
                     if isinstance(view_result, ErrorResponse):
-                        await txn.commit()
+                        await tm.commit(txn=txn)
                     elif isinstance(view_result, UnauthorizedResponse):
-                        await txn.abort()
+                        await tm.abort(txn=txn)
                     else:
-                        await txn.commit()
+                        await tm.commit(txn=txn)
                 except Unauthorized:
-                    await txn.abort()
+                    await tm.abort(txn=txn)
                 except Exception as e:
                     logger.error(
                         "Exception on writing execution",
                         exc_info=e)
-                    await txn.abort()
+                    await tm.abort(txn=txn)
             except (KeyboardInterrupt, MemoryError, SystemExit,
                     asyncio.CancelledError, GeneratorExit, RuntimeError):
                 self._exceptions = True