+++ guillotina/db/orm/base.py
@@ -42,7 +42,7 @@ class BaseObject(object):
     # This slots are NOT going to be on the serialization on the DB
     __slots__ = (
         '__jar', '__oid', '__serial', '__of', '__parent', '__annotations',
-        '__name', '__cache', '__new_marker', '__changes')
+        '__name', '__immutable_cache', '__new_marker', '__changes', '__locked')
 
     def __new__(cls, *args, **kw):
         inst = super(BaseObject, cls).__new__(cls)
@@ -53,9 +53,10 @@ class BaseObject(object):
         _OSA(inst, '_BaseObject__parent', None)
         _OSA(inst, '_BaseObject__name', None)
         _OSA(inst, '_BaseObject__annotations', {})
-        _OSA(inst, '_BaseObject__cache', -1)
+        _OSA(inst, '_BaseObject__immutable_cache', False)
         _OSA(inst, '_BaseObject__new_marker', False)
         _OSA(inst, '_BaseObject__changes', {})
+        _OSA(inst, '_BaseObject__locked', False)
         return inst
 
     def __repr__(self):
@@ -220,21 +221,19 @@ class BaseObject(object):
 
     __annotations__ = property(_get_annotation, _set_annotation, _del_annotation)
 
-    # CACHE
-    # -1 : No cache
-    # 0 : Allways
-    # X : ttl
+    # Immutable cache
+    # if we want to cache something in memory forever, think root db object here
 
     def _get_cache(self):
-        return _OGA(self, '_BaseObject__cache')
+        return _OGA(self, '_BaseObject__immutable_cache')
 
     def _set_cache(self, value):
-        _OSA(self, '_BaseObject__cache', value)
+        _OSA(self, '_BaseObject__immutable_cache', value)
 
     def _del_cache(self):
-        return _OSA(self, '_BaseObject__cache', None)
+        return _OSA(self, '_BaseObject__immutable_cache', False)
 
-    __cache__ = property(_get_cache, _set_cache, _del_cache)
+    __immutable_cache__ = property(_get_cache, _set_cache, _del_cache)
 
     # __changes__:  Identifier of the object.
     def _get_changes(self):
@@ -260,3 +259,16 @@ class BaseObject(object):
         _OSA(self, '_BaseObject__new_marker', False)
 
     __new_marker__ = property(_get_new_marker, _set_new_marker, _del_new_marker)
+
+    # __locked__:  marks an object as being locked for writing
+    # and that after the transaction commits, it should unlock
+    def _get_locked(self):
+        return _OGA(self, '_BaseObject__locked')
+
+    def _set_locked(self, value):
+        _OSA(self, '_BaseObject__locked', value)
+
+    def _del_locked(self):
+        _OSA(self, '_BaseObject__locked', False)
+
+    __locked__ = property(_get_locked, _set_locked, _del_locked)