+++ guillotina/db/storages/base.py
@@ -1,15 +1,16 @@
+
+
 class BaseStorage(object):
 
-    _cache = {}
+    _cache_strategy = 'dummy'
     _read_only = False
     _transaction_strategy = 'merge'
 
-    def __init__(self, read_only=False, transaction_strategy='merge'):
+    def __init__(self, read_only=False, transaction_strategy='merge',
+                 cache_strategy='dummy'):
         self._read_only = read_only
         self._transaction_strategy = transaction_strategy
-
-    def use_cache(self, value):
-        self._cache = value
+        self._cache_strategy = cache_strategy
 
     def read_only(self):
         return self._read_only