+++ guillotina/tests/mocks.py
@@ -9,7 +9,12 @@ class MockDBTransaction:
 
     def __init__(self, storage, trns):
         self._storage = storage
-        self._transaciont = trns
+        self._transacion = trns
+
+
+class MockTransaction:
+    def __init__(self):
+        self._tid = 1
 
 
 @implementer(IStorage)
@@ -18,8 +23,12 @@ class MockStorage:
     _cache = {}
     _read_only = False
     _transaction_strategy = 'merge'
+    _cache_strategy = 'dummy'
+    _options = {}
 
-    def __init__(self):
+    def __init__(self, transaction_strategy='merge', cache_strategy='dummy'):
+        self._transaction_strategy = transaction_strategy
+        self._cache_strategy = cache_strategy
         self._transaction = None
 
     async def get_annotation(self, trns, oid, id):
@@ -36,5 +45,7 @@ class MockStorage:
 class MockTransactionManager:
     _storage = None
 
-    def __init__(self):
-        self._storage = MockStorage()
+    def __init__(self, storage=None):
+        if storage is None:
+            storage = MockStorage()
+        self._storage = storage