+++ guillotina/tests/test_postgres.py
@@ -244,7 +244,8 @@ async def test_should_not_raise_conflict_error_when_editing_diff_data(postgres, 
     await cleanup(aps)
 
 
-async def test_should_raise_conflict_error_when_editing_diff_data_with_resolve_strat(postgres, dummy_request):
+async def test_should_raise_conflict_error_when_editing_diff_data_with_resolve_strat(
+        postgres, dummy_request):
     request = dummy_request  # noqa so magically get_current_request can find
 
     aps = await get_aps('resolve')