+++ CHANGELOG.rst
@@ -1,7 +1,11 @@
 1.0.0a22 (unreleased)
 ---------------------
 
-- Nothing changed yet.
+- Fix QueueUtility to properly get transaction object before working on view
+  [vangheem]
+
+- Update storage caching interfaces to make them easier to use
+  [vangheem]
 
 
 1.0.0a21 (2017-05-09)